+++ src/hooks/api/member/useGetUserSummary.ts
@@ -3,7 +3,7 @@ import { useQuery } from '@tanstack/react-query';
 import { memberApi } from '@/api/memberApis';
 import QUERY_KEYS from '@/constants/queryKeys';
 
-export default function useGetUserSummary() {
+export default function useGetUserSummaryMutation() {
   return useQuery({
     queryKey: [QUERY_KEYS.USER_SUMMARY],
     queryFn: () => memberApi.getUserSummary(),
